refactor(wave-2): fix scrutinizer blockers — file size, composite query, SQL cache
Addresses 6 blockers from Wave 2 quality review:

1. Split temporal/storage/mod.rs (568 lines) into storage/mod.rs +
   storage/query.rs + storage/build_ops.rs (each < 300 lines)

2. Split cmd/search/mod.rs (484 lines) into mod.rs + dispatch.rs. The
   top-level run() is now a thin router; the three query dispatch paths
   (standalone blast-radius, standalone temporal, lexical+composite)
   live in dispatch.rs

3. Fix biased candidate pool in multi-signal standalone temporal queries.
   --hot --risky previously fetched from hot only; now unions candidates
   from all enabled signals before rerank

4. Use prepare_cached in load_score_for to amortize SQL preparation cost
   across rerank iterations (was N×M fresh prepared statements, now 1
   cached prepare per signal kind)

5. Add end-to-end CLI integration tests for temporal flags:
   --build-temporal, --hot standalone, --blast-radius conflicts,
   --hot --cold warning

6. Add clippy::panic to temporal_test_helpers allow list for consistency
